# public/openapi.yaml
openapi: 3.0.3
info:
  title: 탄소이음 EMS API
  description: |
    탄소이음(Carboneum) 에너지 관리 시스템 공개 API.
    멀티테넌트 에너지 데이터 수집, 탄소 배출량 계산, 실시간 모니터링을 제공합니다.

    ## 인증
    모든 API 요청에는 `Authorization: Bearer {API_KEY}` 헤더가 필요합니다.
    API 키는 **설정 → API 키 관리** 메뉴에서 발급받을 수 있습니다.

    ## Rate Limit
    | 플랜 | 한도 |
    |------|------|
    | Trial | 100회/시간 |
    | Basic | 500회/시간 |
    | Pro | 2,000회/시간 |
    | Enterprise | 10,000회/시간 |

    Rate Limit 초과 시 `429 Too Many Requests` 응답이 반환됩니다.
    응답 헤더 `X-RateLimit-Remaining`, `X-RateLimit-Reset`으로 잔여 횟수를 확인하세요.
  version: 1.0.0
  contact:
    name: 탄소이음 기술지원
    email: carbonieum.official@gmail.com
    url: https://carboneum.kr/support
  license:
    name: 상업용 라이선스
    url: https://carboneum.kr/legal/terms
servers:
  - url: https://carboneum.kr/api
    description: 프로덕션 서버
  - url: http://localhost:3000/api
    description: 로컬 개발 서버
security:
  - BearerAuth: []
...
paths:
...
  /ai/optimize:
    post:
      tags: [AI]
      summary: 에너지 최적화 추천
      description: 소비 패턴을 분석하여 에너지 절감 추천 규칙을 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [siteId]
              properties:
                siteId:
                  type: string
      responses:
        '200':
          description: 추천 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizeResponse'

  # ─────────────────────────────────────────────────────
  # 게이트웨이
  # ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API 키 (설정 → API 키 관리에서 발급)
    GatewayToken:
      type: apiKey
      in: header
      name: X-Gateway-Token
      description: 게이트웨이 토큰 ({serial}:{secret} 형식)
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        code:
          type: string
...
    OptimizeResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                recommendations:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [peak_shift, standby_reduction, weekend_schedule, high_variance]
                      title:
                        type: string
                      description:
                        type: string
                      estimatedSavings:
                        type: number
                        description: 예상 절감액 (원/월)
                      priority:
                        type: string
                        enum: [high, medium, low]
...